Split webscrambles: Second cloudscrambles subproject
