+++ cloudscrambles/src/main/resources/application.conf
@@ -5,7 +5,7 @@ ktor {
 
     application {
         modules = [
-            org.worldcubeassociation.tnoodle.server.webscrambles.WebscramblesServer.spinUpLocally
+            org.worldcubeassociation.tnoodle.server.cloudscrambles.CloudscramblesServer.spinUp
         ]
     }
 }
