+++ cloudscrambles/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="ISO-8859-1" ?>
+<web-app xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" version="3.0">
+    <servlet>
+        <display-name>KtorServlet</display-name>
+        <servlet-name>KtorServlet</servlet-name>
+        <servlet-class>io.ktor.server.servlet.ServletApplicationEngine</servlet-class>
+        <!-- path to application.conf file, required -->
+        <init-param>
+            <param-name>io.ktor.config</param-name>
+            <param-value>application.conf</param-value>
+        </init-param>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>KtorServlet</servlet-name>
+        <url-pattern>/</url-pattern>
+    </servlet-mapping>
+</web-app>
