+++ server-ktor/build.gradle.kts
@@ -1,18 +1,13 @@
 import configurations.CompilerSettings.KOTLIN_JVM_TARGET
 import configurations.Languages.attachRemoteRepositories
 
-import dependencies.Libraries.APPLEJAVAEXTENSIONS
-import dependencies.Libraries.GOOGLE_CLOUD_STORAGE
 import dependencies.Libraries.JUNIT_JUPITER_API
 import dependencies.Libraries.JUNIT_JUPITER_ENGINE
-import dependencies.Libraries.KOTLIN_ARGPARSER
 import dependencies.Libraries.KTOR_GSON
-import dependencies.Libraries.KTOR_SERVER_CIO
 import dependencies.Libraries.KTOR_SERVER_NETTY
 import dependencies.Libraries.KTOR_SERVER_HOST_COMMON
 import dependencies.Libraries.KTOR_SERVER_SERVLET
 import dependencies.Libraries.LOGBACK_CLASSIC
-import dependencies.Libraries.NATIVE_TRAY_ADAPTER
 import dependencies.Libraries.MARKDOWNJ_CORE
 
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
@@ -26,7 +21,6 @@ plugins {
 }
 
 dependencies {
-    api(KTOR_SERVER_CIO)
     api(KTOR_SERVER_NETTY)
     api(KTOR_GSON)
 
@@ -38,11 +32,7 @@ dependencies {
 
     implementation(KTOR_SERVER_SERVLET)
 
-    implementation(APPLEJAVAEXTENSIONS)
-    implementation(NATIVE_TRAY_ADAPTER)
     implementation(MARKDOWNJ_CORE)
-
-    implementation(GOOGLE_CLOUD_STORAGE)
 
     testImplementation(JUNIT_JUPITER_API)
     testRuntime(JUNIT_JUPITER_ENGINE)
